Commited an example of static and final keywords. Was bored during this class so I messed around a little. Also played with a JavaFX2 example with playing Mp3's
Currently working on a text based game that will simulate a game of craps
there can be multiple players and there will be sounds played at certain
moments. There is also background music. This is done with an external
library (javazoom) which is from the princeton university intro to CS library.